requirements.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.3
idna==2.10
oauthlib==3.1.1
pycodestyle==2.7.0
PySocks==1.7.1
python-dotenv==0.19.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tuya-bulb-control==0.1b3
tweepy==3.10.0
urllib3==1.26.2
